Weapon patch
Changes:
Remake weapon take-throw system. Maked singleton with ID's and PackedScene's dictionary. Maybe I will change ID's on weapon Name.
Weapon now is area2D and now it in Weapons Array in player script.
When we throw weapon we add that PackedScene and hide weapon in player node.
When we take weapon, we check visibility of this weapon and name. And now we can take this.
Add new buttons in pauseMenu - Load and Save

Bugs:
Load aren't working now currectly.